LayoutTests/platform/mac-leopard/media/controls-drag-timebar-expected.txt

Test that drag the timebar thumb causes seeks.

RUN(video.autoplay = true)
RUN(video.src = 'content/test.mp4')
EVENT(playing)
EVENT(seeked)
Time: 0.3
EVENT(seeked)
Time: 1.1
EVENT(seeked)
Time: 1.9
END OF TEST
